Add a checkstyle rule that rejects uses of spaces for indents
It seems we've been accumulating spaces in a few cases, this will
make sure we're consistent going forward.

(Another commit will fix up the existing issues.)

This is mostly taken from https://stackoverflow.com/a/28550141/292166

I changed the regexp so it only triggers for 4 spaces, as there's a
few instances where we use 1-3 spaces for alignment.

part of #1161