FAPI2 ConnectID: Allow overriding of scope so we can add tests to CI
The ConnectID tests use only the 'openid' scope, but expect specific
fapi2sp/ms behaviours. This makes it hard for Authlete to configure the
server they provide for us to use in our CI to pass these tests,
as the interactions for some of the negative tests look
exactly like those expected to pass in the OpenID Connect Core
tests.

To solve this, we add a way to override the scope used in the idmvp
tests so that our test config for this test can specify an extra scope
that triggers the necessary fapi2 behaviours.