Remove $DEV_MACHINE$ from .idea
It's still not clear why IntelliJ started adding these; I've installed
the latest IntelliJ on a fresh machine and it seems to have stopped.

I can only assume it was some kind of bug in one particular IntelliJ
version or a plugin or something, resulting in DEV_MACHINE set to true
being added to my path variables as true. And it seems once that is added,
IntelliJ will replace any instance of 'true' in the config file with
the path variable instead, making the whole thing self-perpetuating.

Hence probably everyone will need to remove DEV_MACHINE from:

settings -> appearance & behaviour -> path variables

to stop it doing this.