Allow CI to pre-provision tokens
It's useful if we can set tokens into each review app on creation so
that we can use the same tokens everywhere.

Read in a environment variable set in the gitlab CI settings and apply
tokens from it into mongo.

The environment variable can have multiple tokens seperated by '|' and
the contents is the mongo document, but with a matcher before it so
it is suitable for use in a .update() call.

The documents can be created using the suite and then queried from
mongo.