Query Supply Only Required Parameters To Authorization Endpoint
A new test, EnsureAuthorizationRequestContainsOnlyValidParamsWhenUsingPAR,
has been added to validate the parameters sent to the authorization
endpoint when using PAR.

The CI tests test this by running the OP test against the RP test. This
commit updates the tests sending the authorization requests in this
scenario to supply only the required query parameters.