Fix False Warnings When Validating ID Token Claims
The EnsureIdTokenDoesNotContainNonRequestedClaims condition checks
the claims received in the ID token against a fixed list of valid claims.

This commit modifies this logic to allow for the fact that OpenID Connect
allows claims to be specified via scopes in the authorization request.

The condition now, in the OpenID Connect case, converts scopes in the
authorization request into a set of claims and adds them to the valid
claims list thus allowing them to be contained in the Id Token.