OP/RP: Allow Unknown Value In Authorization Request Object
Add an expected failure to cover the case where we are deliberately
sending an unknown value in the authorization request object.

The fapi-au-connectid-rp-test-config.json configuration file has
been duplicated. The test requiring the exception has been updated
to use this duplicate giving it a unique signature for use in the
expected failure file.