Guard against missing runningTestInformation element
According to https://gitlab.com/openid/conformance-suite/-/blob/master/src/main/resources/static/templates/logHeader.html#L104
the `runningTestInformation` element is only added to the document if it's
not a public log. This commit avoids calling the `pinVisitSectionOnTopWindow`
function if it's a non-public log as well as increasing defensiveness in
the function itself.