Add extra flexibility around http response when TLS cert missing
We already allow a range of responses when a token endpoint call is
missing a TLS certificate, but we currently require a http 400 or 401,
but following the logic we already have a 403 response would be fine too.

Change the test to require a 4xx code, but only warning if it's not a
400/401, in the same way we warn about non-json content.

To avoid creating more 'checkhttpstatusfromXXXXisYYYY' functions, the
token endpoint calls are updated to return the same full object that
other http calls return.

For consistency, apply the same change to FAPI1, FAPI2 and FAPI-CIBA.

closes #1225