Fix second client test so it uses 128 character state
This got broken in a refactor at some point, in multiple ways:

The exec() command wasn't being call()ed, which in turn revealed that
putInteger was returning void rather than Command as other parts of
this builder do. Then when retrieving the value with getInteger an
exception was thrown because null (in particular JsonNull) had been
stored into the environment. Add a version of removeNativeValue to
avoid that.

Also did a search for all exec() calls without a call() and found two other
instances.

Spotted whilst testing the change for #1226