Add message warning that webfinger is deprecated
Migrating to newer versions of SpringBoot will almost certainly mean
moving from the org.mitre openid client to the spring security one, and
the latter does not currently support webfinger. (The current org.mitre
openid client is incompatible with changes in the latest spring versions.)

When this happens users will not be able to access their old test results,
so advanced warning is necessary.