Fix for regression in CallTokenEndpointAllowingTLSFailure
Non-JSON responses were throwing an exception as I'd omitted to add
the flag to ignore json parse failures. This wasn't caught in the CI
as the systems we test against all have valid responses - so add a
unit test for CallTokenEndpointAllowingTLSFailure as well.

It's based on CallTokenEndpoint_UnitTest but with updates to test this
particular code path plus some others.