Cleanup display of TestFailureException
Failures like "Got an HTTP response we weren't expecting" were showing
the user a stacktrace, which is unhelpful and might make the user
think it's a bug on our side.

Also display the path in the error, so it's clearer what endpoint
was called without going hunting for a different log entry.