Upgrade to SpringBoot 2.5
Or in particular, 2.5.15, which is currently the latest release in the
2.5 series.

getErrorPath() in the error controller is removed as this method is no
longer present in the interface.

As per the docs the location is now set using server.error.path, which
as we haven't set it defaults to /error, the value we currently expect:

https://github.com/spring-projects/spring-boot/blob/v1.3.0.RC1/spring-boot-autoconfigure/src/main/java/org/springframework/boot/autoconfigure/web/ErrorProperties.java#L33

To verify the error controller was working correctly I tried:

curl -k https://localhost:8443/api/runner

and verified a breakpoint in FAPIErrorController was hit and that it
correctly resulted in JSON being returned as expected.

I read through the release notes and did not spot anything scary:

https://github.com/spring-projects/spring-boot/wiki/Spring-Boot-2.5-Release-Notes

This may mean we can now move to java 16 (springboot 2.5 adds official
compatibility with java 16) but I have not tried that yet.

closes #1230