Upgrade to SpringBoot 2.6
This doesn't seem to cause any obvious issues (beyond the small change to
WebSecurityConfiguration.java made in the previous commit).

The release notes are here:

https://github.com/spring-projects/spring-boot/wiki/Spring-Boot-2.6-Release-Notes

As best as I can tell the PathMatcher changes don't affect us as we don't
use mvcMatchers, and also the path matchers we do have seem to have the
required leading "/" mentioned in the release notes.