FAPI2 OP: Improve token type case insensitivity test
Add a description and fix some confusing phrases - the thing we invert the
case of is the token type returned from the token endpoint, i.e the one
we include in the authorization header sent to the resource endpoint.

Also rename conditions and change it to a 'stopOnFailure' (as it makes
no sense for the test to continue if it couldn't invert the token, and it
really shouldn't ever fail).

Changed as a result of a support query confused as to the why the token
should be treated as case insensitive.