Change token endpoint calls to 'stopOnFailure'
I can't see any situation where it can make sense to carry on after
we get a failure calling the token endpoint.

Changed as a result of a log which looked confusing as the token
endpoint call had failed but the test had continued on, resulting
in CheckTokenEndpointReturnedJsonContentType failing one of it's
@PreEnvironment conditions and showing the "this might be a bug
in the test suite" message.