Fix runtime exception not being displayed helpfully in the frontend
Errors like the RuntimeExceptions thrown when the user tries to create
a testplan that has invalid variants for a particular FAPI profile are
not being displayed to the user anymore.

The format these errors are returned in must have changed in one of the
SpringBoot upgrades, as I'm sure errors like this used to get displayed
fine.

At least for now, pick where the error is returned out of the
response.