FAPI RP Tests: Generate Mixed Case FAPI Interaction Id's
The FAPI1 OP tests use ValidateFAPIInteractionIdInResourceRequest to
validate that the interaction id is a valid UUID.

This is done using UUID.fromString() which accepts both upper and
lower case hex digits in the UIUD string.

This commit extends the usage of ValidateFAPIInteractionIdInResourceRequest
to the FAPI2 OP tests and updates CreateRandomFAPIInteractionId to generate
mixed case UUID strings ensuring the RP tests have the same coverage as the
OP tests.