Fix issue in compare job when new test job is added
The compare job fails when unpacking the zip files if a passing result
for that job doesn't exist on master yet, as curl ends up writing out
the 404 html as the zip file.

Add --fail so that curl exits with failure for http errors, and skip
unpacking the zip when that happens.