Exclude fapi1-advanced-final-client-brazildcr-happypath-test from client CI
We've never been able to run this, and previously weren't running it as
it was only listed in the Brazil specific plan - now it's part of the
main plan we have to explicitly exclude it.

This requires a few extra changes to run-test-plan because the current
exclusion list didn't work for all types of tests, and we needed to be
able to override it to run this test module in the RP-against-OP tests.