UI: Fix Invisible 'Status' In Logs View
It was noted that the status result in every 2nd row in the 'logsListing'
table was invisible (white on white).

The root cause is that 'testStatus-*' styles are applied to these cells.

These styles are designed for displaying test status in the plan/log-details
views and are not applicable here. Remove them.