FAPI1Adv RP: Update generic checking of the acr claim
It seems this hadn't been updated to match the final version of the FAPI1
specification - in particular the spec previously said that an essential
acr claim had to be made, but this was removed in the final version.

Update the code so that it checks the validity of any acr claim that is
present, but does not error if the acr claim is missing, or is not
essential, or does not request values.