Improve error reporting when JWT payload is invalid JSON
Unfortunately com.nimbusds.jose.Payload's toJSONObject catches parse
exceptions and simply returns null so we can't see the actual underlying
error message. We extract some of the code from nimbus so that the detail
of the parse exception can be shown to the user.

The code is not efficient (and has never been) as it does multiple
conversions to/from JSON,  however as per the comment added in the code
making it more efficient loses some checks so I have left it as is for
now, however we might want to address that at some point. It would also
be helpful if we could throw test condition errors from JWTUtil so we
could report errors in a better way - but that would require a larger
refactor.

Helpful for diagnosing things like #1284