Further updates to the Brazil OF spec links
I'd forgotten to update LogEntryHelper for some of the recent changes,
and the DCR specs we linked to have both now been deleted so replace
them with the latest version.

Note that there is no change log for the DCR specs, and the old versions
are no longer available, so I have no idea if there are normative changes
that might need updates to the tests.