Improve error message when we find an unexpected type in the Environment store
As per #1315 the error message doesn't make it very clear that the value not
being present is an acceptable outcome too. Try to improve this in the generic
case, there is an argument for adding more explicit checking in ValidateIdTokenNonce
instead but I'm reluctant to do that right now as this case shouldn't occur often.

closes #1315