OID4VP: Initial support for testing via Browser API preview
As per:

https://github.com/WICG/digital-credentials?tab=readme-ov-file

and:

https://openid.net/specs/openid-4-verifiable-presentations-1_0-21.html#name-openid4vp-profile-for-the-w

A new request type is added to BrowserControl that ends up communicating
to log-detail.html a request to be sent to the browser API, and a url for
sending the result of the API call back to.

The data returned by log-detail.html is parsed and unpacked into the same
authorization_endpoint_response object in the environment that the
non-browser-API response ends up in, and the same validation code then
run on the response.

This is a bit rough, for example the 'Process with test' button for
browser API is never moved to the 'visited' list like a normal redirect
is. I've also not exhaustively verified that the requests actually align
with the protocol, and have only been able to test unsigned requests.
Only non-encrypted responses are supported.

Tested on an Android phone with the wallet linked from here:

https://github.com/WICG/digital-credentials/wiki/HOWTO%3A-Try-the-Prototype-API-in-Chrome-Android

but with the caveat that a few hacks were necessary to make it work with
that wallet (the hacks aren't included in this commit).

Thanks to Helen Qin, Lee Campbell, Rick Byers for help getting an environment
setup to test this, and to Filip Skokan for help with some of the javascript.