Fix random diff in the compare_cloud job
The CIBA OP-against-RP tests run the
fapi-ciba-id1-ensure-other-scope-order-succeeds test module twice with the
same config, once against fapi-ciba-id1-client-test and once against
fapi-ciba-id1-client-refresh-token-test.

This means that we end up with two runs that have the same testplan /
variant / test module / description but that produce two different logs,
and it's currently random which one is used in the comparison depending
which order the file system stores the files I think.

Add a second configuration file with a different description so that the
result for the refresh token run is kept separate to the normal happy
path run.