RPInitiatedLogout: Better error if multiple calls to post_logout_redirect
A user got fairly confused because if you call the post_logout_redirect
url multiple times then you end up with the next stage of the test
running multiple times and then weird things happen.

Just set a flag when we initiate a logout, clear it when we get the
callback and error if we get a callback when the flag isn't set.