Reduce the SSD size for review app environments
SSD currently makes up a good third of the costs of our cloud environments,
and with the recent changes to increase the number of available environments
this could get worse.

I think most review apps don't come close to needing 25G of mongodb storage,
so reduce down to 3G which I think'll be more than enough in most cases.

It's always possible to edit this to increase it on the relevant branch
if more is needed.