Fix for user's old results not showing up
The 'principal' field after the SpringBoot 3 upgrade had an extra field
and was in a different order, resulting in the mongodb queries not
returning the user's old data.

Return it to the previous layout.

Tested on 'demo' and verified that only with this fix could I see some
of the old results from my google account.