Further fixes for gitlab deployments
We had not set a KUBE_NAMESPACE, which previously gitlab was setting
automatically for us at least when using the certificate based integrations
and possibly also when we were using the gitlab agent but had the
certificate based integrations still enabled but not actually being
used.

The missing KUBE_NAMESPACE results in the agent/autodevops deploying
to a namespace called 'default' but then our own kubectl calls tried
to use the namespace '' and failed in various ways, initially when
trying to figure out if the mongodb pod had started or not.

So now we explicitly set KUBE_NAMESPACE in .gitlab-ci.yml. The namespace
picked should match the one gitlab was previously using, which may mean
any older branches that get rebase will still use the same namespace as
they were.

Additionally the 'stop_review' jobs were not calling
"auto-deploy use_kube_context" which meant they didn't initialise the
gitlab agent and hence were failing once the certificate based integration
was disabled.