Fix external url override mechanism after split of mtls hostnames
The refactoring hadn't included adding the TestDispatcher path prefix
(/test/) into the override url, resulting it in being invalid.

Obviously this won't work for /test-mtls/ paths but I don't think it ever
did, we can try and solve that if/when it becomes an issue.

This is sufficient for OID4VP wallet testing to work locally again.