OID4VP: Include x5c if present for non-mdl formats
Currently we only include the x5c header in the signed request jwt if
the credential format is set to mdl, which was completely wrong as it
should have actually depended on if the client_id_scheme was set to one
that used x5c certificates.

Update that, but also change the code so even in the non-x509-based
authentication schemes if the key has an x5c we include it.