Fix bad variant selections not showing proper error message
I presume the recent-ish upgrades to SpringBoot have changed the format
we are getting errors back, so that the JSON we get back when exceptions
are thrown no longer has an outer 'error' object. Update the code that
parses them for this (and tidy it up a little).

(I'm not convinced the other catch paths here can actually be triggered,
but haven't touched them for now.)