Integrate google's errorprone checker
See https://errorprone.info for background

Integrated as per the maven instructions here:

https://errorprone.info/docs/installation

(including the creation of a .mvn/jvm.config to add the required extra
compiler flags)

I initially looked at this because I wanted something to check that
switches on enums had all the values listed.

However it did flag a few other things including what definitely looks
like a bug in FAPIValidateIdTokenEncryptionAlg where it wouldn't have
actually failed.

There's a few comment themes - mostly deleting useless javadoc comments
and changing LinkedList to ArrayList.

As we've done with previous checkers, the configuration of excluded
checks is a bit haphazard and I've disabled a few checks solely on the
basis we have a non-trivial amount of code that fails them so they're
probably not too crucial.

The unit tests are excluded are they produced a ton of warnings too.