OID4VP: Allow a client_id to be specified when using x509_san_dns
If the certificate being used is one that the wallet places a higher
level of trust in, then the client can use any response_uri (rather
than needing to use one that matches the client_id and hostname in the
certificate) - so in this case people might want to use a certificate
that doesn't have demo.certification.openid.net in it so let them
specify which client_id they do want to use.

I did wonder about extracting the SAN entry from the certificate and
automatically using that as the hostname, but the certificate can have
multiple of these so it's better to let the user specify exactly which
one they want to use.

The old behaviour of using the conformance suite hostname as the client_id
is preserved if the client_id is left unset in the configuration.

(Requested by Kristina for German wallet testing.)