Update AbstractSignJWT To Respect Single Element 'aud' Array Claims
By default instatiating the 'SignedJWT' class will convert a single element
'aud' array claim to a string. This is done by JWTClaimsSet::toJSONObject()
called when converting a JWTClaimsSet to a Payload in the SignedJWT
constructor for passing to the JWSObject constructor.

NB. Instantiating JWSObject encodes the payload for use when signing.
Modifying the payload after this point does not re-encode it and thus
the new payload is not included in the signed data.

Here we add a new method, performSigningEnsureAudIsArray(), which adds the
ability to maintain a single element 'aud' array claim as an array.

This is achieved by manually generating a payload with a single element
'aud' array claim and instantiating JWSObject directly.