Upgrade to latest nimbus-jose-jwt
Mainly done to pull in the fix for this issue:

https://bitbucket.org/connect2id/nimbus-jose-jwt/issues/565/x509certchainutils-parse-method-makes-it

so that our users get better diagnostics.

Full changelog is here, I didn't spot anything that looked like it might
be problematic for us:

https://bitbucket.org/connect2id/nimbus-jose-jwt/src/master/CHANGELOG.txt

Add unit test is also added for an error that happened against the
brazil mock bank when trying to upgrade to 9.41.1, but turned out to be
solved in 9.41.2.