OID4VP: Refactor adding encryption key to client_metadata
It's needed for any case encryption is used (which as we currently assume
a jwt response means 'encrypted only' means anytime the jwt response is
used).

Configuration parameters are added for the enc & alg values - the mdl spec
is pretty clear about what values to use, but SD-JWT / HAIP etc don't seem
to give any guidance currently.

This fixes the jwt response mode for sd-jwt, as previously it wasn't
sending the encryption key.