FAPI (PAR): Ensure The 'request_uri' Can Be Reused Prior To Authentication Completion
As per https://bitbucket.org/openid/fapi/issues/635/one-time-use-of-request_uri-causing-error

Tests have been added to issue a GET request, including the 'request_uri', to
the authentication endpoint prior to executing a successful authentication flow.

This should succeed.

NB. The HttpClient can now be configured to handle redirection. This is enabled
    in the new tests.