log-detail: Fix incorrect block names being shown
The block name is only still valid if there's actually a block id present,
the lack of this check was causing some failures to show as being
associated with client deregistration when that block had already
ended.

This probably still isn't quite right as it should actually be looking
up the block name from the blockid, but it's better than it was.