OID4VP: Allow other mdl profiles to be run
Although ISO 18013-7 is very prespective, people are using mdls in other cases
and with the browser API, so it's not helpful for us to stop them running
tests for combinations they do want to test.

Remove all the cases where we block creation of the test plan, and fill
in the cert profile so it's slightly more obvious what is being tested.