Reduce number of CheckDistinctKeyIdValueInServerJWKs warnings in CI
The fapi2 client tests were recently changed to have duplicate keyids
in order to check key selection - unfortunately this results in a
warning in every rp-against-op fapi2 test, and we've recently added
more of these (so the CI is currently failing, as I merged something
without waiting for the CI to finish after a rebase).

Move this warning into only the happyflow test so that it's less noisy.