Initial version of OID4VP verifier tests
The modifies the test modules created from the OpenID Connect client tests
in the previous commit to actually test OID4VP.

Instructions for running these tests are found here:

https://openid.net/certification/conformance-testing-for-openid-for-verifiable-presentations/#verifier

The initial version is really quite rough with quite a few checks
and options missing. However it does work and has been tested against
both Animo's verifier ( https://funke.animo.id ) and in the op-against-rp
tests. I think there's still some code left that's not applicable to
OID4VP and quite a bit of refactoring is needed on some of the new code.

The script used to generate the self-signed x509 certificate used in the
test configurations can be found in:

conformance-suite-private/test-keys-for-vp-tests

Future work will fill some of the gaps and start to add negative tests,
more flexibility, etc.

part of #1411