OID4VP: Update identity-credential library used for mdoc parsing
Animo reported they were getting an error like this:

2024-11-02 18.14.50.908: DEBUG: DeviceResponse: issuerSignedAuthenticated: true
Exception in thread "main" java.lang.IllegalArgumentException: Item 2 (payload) is a simple but not the value null
        at com.android.identity.internal.Util.coseSign1CheckSignature(Util.java:726)
        at com.android.identity.mdoc.response.DeviceResponseParser$DeviceResponse.parseDeviceSigned(DeviceResponseParser.java:287)
        at com.android.identity.mdoc.response.DeviceResponseParser$DeviceResponse.parse(DeviceResponseParser.java:360)
        at com.android.identity.mdoc.response.DeviceResponseParser.parse(DeviceResponseParser.java:136)

and that upgrading to the later library helped this.

Unfortunately upgrading is not entirely straightforward as the library
has been split up into multiple packages and some of the APIs changed.

Also cleanup some of the commented out code & add a unit test with the
mdoc credential animo were using.