FAPI2SPID2 OP: Rework how client assertion auds are tested
Currently the test defaults to sending the token endpoint, which then
results in the op-against-rp tests failing as the previous commit
changes them to require the iss as the aud as per FAPI2SPID2.

Now the tests send iss in most tests, and the test that previously
checked that iss was accepted as the aud has been changed to check
that the token endpoint is accepted, as per the text in FAPISPID2.

part of #1400