ValidateRequestObjectSignature: Fix logging in error case
Sending a List<JWK> to args() has always been problematic when the JWKS
contains x5c entries - this happened again here, so do a proper conversion
into a JWKS instead.

The problem was seen when running the OID4VP Wallet tests against animo
with a bad configuration.

Also include same logging in the success case as that means it's more
likely we'll see any problem.