Refactor CallTokenEndpointAndReturnFullResponse
It has some generic properties that mean we should reuse the code, so
split out the base layer.

I've not changed anything else to use the new base class, but do have some
in-progress code for testing verifiers for OID4VP that uses the new base
class to call the direct post endpoint.

The only deliberate change in behaviour is that the new class supports
JSON objects as well as simple string values, as this was required to
support the presentation_submission parameter in OID4VP's response.