OID4VP: Rename all things to consistently include 'verifier'/'wallet'
Sort out some inconsistencies in naming by making sure test modules,
test plans, and their names all consistently have either 'verifier'
or 'wallet' in them, and also include 'id2' and 'oid4vp' somewhere.

This is mostly advance preparation in the knowledge we'll need to create
ID3 tests in the next month or two.

A few of the needed spec references are missing from ID2 - e.g. SD-JWT
ones; this is because the text around SD-JWT was added to the spec after
OID4VPID2. Because of this we link to the relevant text in what will
become ID3 instead.