Adjust index.html language around 'View all plans' etc
The 'all' is confusing when the user is looking to view their own test plans.
This way is possibly slightly confusing for admins, but this part of the html
is not currently easy to vary for admin vs non-admins.

Also tweak the layout a little.

Based on feedback from Kristina