Add possible workaround for #1458
As per info on the issue, the problem seems to be that the xmlhttprequest
POST that sends the result back to the conformance suite fails because
sometimes it happens whilst nginx is restarting.

We've never logged failures here so start doing so.

Also if we do get an error, try again after 2 seconds.

I think this won't make us retry after a http level error (only a tcp level
one); I'm not sure if we should/need to do that.