FAPI1 OP Tests: Commonise processParResponse()
Create a common implementation of processParResponse() in a sub-class of
AbstractFAPI1AdvancedFinalExpectingAuthorizationEndpointPlaceholderOrCallback().

This implementation calls processParErrorResponse() which must be
implemented by derived classes to handle the PAR response error flow.

Tests that currently implement processParErrorResponse() and sub-class
AbstractFAPI1AdvancedFinalExpectingAuthorizationEndpointPlaceholderOrCallback()
are updated to use this new class and implement processParErrorResponse()
rather than processParResponse().