OID4VP ID3: Change client_id_scheme from parameter to part of client_id
As per changes in the draft. This has the effect of making sure that
"aud" in the sd-jwt kb jwt etc are also checked to include the
client_id_scheme in the client_id.

An 'orig_client_id' is added to the root of the environment that is the
client_id without the scheme, including in the id2 tests for consistency.