Fix run-tests.sh command truncation on gitlab ci
The 'xargs' on the gitlab CI is the busy box one which defaults to a 32
kilobytes-ish buffer:

https://github.com/josefbacik/busybox/blob/2332cee00492b69b92ff04cb8ce3207092fd0879/findutils/xargs.c#L538

which we've just hit with the changes in this MR to add the VP ID3 tests to
the CI, so the xargs was I think attempting to split the test list across
2 runs of run-test-plan.py, but as the first one failed the second one never
happened so it wasn't obvious what was going on.

The reason we're using the 'xargs' is that some of the variant names contain
a space (e.g. "[response_type=code\ token]"). I have a vague feeling we adding
a change somewhere so we could actually use a different character as a proxy
for spaces but I can't find it (maybe we should do something like that).

Fix this for now by increasing the buffer size.

Also add a bit of console output that might've made this more obvious...