OID4VP Wallet & Verifier: Add DCQL support
This is a new query language introduced as an alternative to presentation
exchange in ID3 of VP.

For wallet tests, introduced a new variant to pick the query language, new
configuration entry that takes DCQL instead of PE, packs it into the
appropriate parameter, then when unpack supports the new format returns in
vp_token.

For the verifier tests, a new variant is also added, they look for the new
parameter, and return the new format vp_token. To accomplish this a bunch of
conditions that were directly reading/writing a vp_token entry in the
environment have been changed to instead read/write a "credential"
environment entry.

OP-against-RP tests have been added to verify self-consistency (and helped me
make sure the code at least semi works), but I've not yet been able to test
these against a wallet as the wallet implementers are also only just starting
work on this.

As with the existing PE tests, we do very little processing of the query - as
with PE, the verifier tests have to process the DCQL but only to get the
relevant id to use when returning the credential. We probably want to do more
in the future.