OID4VP Wallet: Improve/fix w3c browser Digital Credentials API support
The browser API tests haven't been used much and only one particular case
(signed requests + encrypted responses) seemed to worked.

There is an awkward tension as we usually target implementer's drafts for
testing, but the current ID3 (draft -23):

https://openid.net/specs/openid-4-verifiable-presentations-1_0-ID3.html

doesn't contain some changes/additions around mdl that are in -24:

https://openid.net/specs/openid-4-verifiable-presentations-1_0-24.html

but that we do want to test at an upcoming interop event. So although I've
left the tests named as ID3 they will start to include elements of the -24
draft...