OID4P ID3: Update test plans names to mention draft24
As per previous commits we've added support for the mdoc sessiontranscript
that is only present in draft 24 so explicitly acknowledge that.

Also explicitly reference -24 in the spec links.

This is kind of awkward as we're blurring the line inbetween ID3 and -24,
but -24 contains some essential changes that we have to incorporate in the
tests, and other groups like the EU are also aligning around -24.