Fix issue where QR codes were sometimes not displayed
Turns out there are different settings for correction level and
amount of data in QR codes. The settings in this MR should allow
for URLs up to 1274 characters.

The `version` can be bumped even more, up to 40, but there are drawbacks
and we might end up with a QR code with so much data that the rendering
area isn't large enough to display it in a way that it can be scanned.