Avoid premature test failure in oidcc-request-uri-unsigned-supported-correctly-or-rejected-as-unsupported test
Previously, we immediately failed the tests if we detect that
the server config reports `request_uri_parameter_supported: false`, which prematurely ended the test execution.

We now skip that hard check and let the test proceed since we later check for the correct `request_uri_not_supported` error anyway.

Fixes #1497

Signed-off-by: Thomas Darimont <thomas.darimont@oidf.org>