OID4VCI Wallet: Handle /.well-known/oauth-authorization-server/
Because RFC8414 requires that .well-known is always at the server root,
this requires changes to TestDispatcher to pass the new call down to the
correct test module.

Some clients (at least Filip's that we use in client_test for oidcc) query
this endpoint in tests that don't currently support it. We preserve the
current behaviour of the existing non-VCI tests by returning a 404 with a
bit of information about the error.