FAPI2 Final OP/RP: Clock Skew
FAPI2 Final mandates that JWTs with an iat/nbf of 10 seconds in the future
are accepted and those with a timestamp of greater than 60 seconds are
rejected.

This was already checked for, but the specification references needed
updating.

NB. A value of 8 seconds in the future is actually applied to allow for
    the time taken to transmit the data.