OID4VP mdoc: Fix issue with macced mdoc
The current code only worked for signed mdoc credentials, but it is legal
to do a MAC operation instead of a signature.

Add support for that by passing the key used for decryption (which is needed
to do the macing) into the identity credential library.

Thanks to David Zeuthen from Google for help with converting the key into
the right kind of object.

Fixes the issue seen in this test log:

https://demo.certification.openid.net/log-detail.html?log=o8LpbtTckcbWE9x