Update files copied from android identity credential library
There have been various changes that mean we need to update to a newer
version, and also don't need some of the files (e.g. the sample mdoc is
now available in a package).

Import the unchanged files from identity-credential - this won't compile,
but I wanted a record of exactly what we imported.

These are from:

https://github.com/openwallet-foundation-labs/identity-credential/

in particular at git sha 2e6d84945074b7c7eecd7eaa180ad040a55f978b, which
was the latest as of end of Wed 9th April 2025.