OID4VCI Issuer: Remove need for a server jwks + remove openid support
One of the test users reported that test was failed because their AS
didn't have a jwks_uri. I can't currently see a reason why one of those
is needed for VCI, so remove that requirement.

At the same time, remove the configuration for plain_oauth vs openid
(there's no need for openid in VCI).

I thought about removing mtls sender constrain support too (it's unlikely
to be useful in the wallet world) but decided against it as I have a
suspicion we might want to support some form of non-dpop token so having
the intrastructure will likely be useful.

Also try to make the VCI part of run-tests.sh more readable.