FAPI2 RP: Exclude jwks_uri when it's unneeded
Server keys are only necessary when using JARM or OpenID, so exclude it
in other places.

This updates the generic "GenerateServerConfiguration"  conditions
(except for the one used only in the OpenIDConnect tests) to not create
a jwks_uri, and adds a new condition to add one in all cases except
when jarm and openid are both not in use in fapi2 final rp.

This originally came out of feedback on the VCI tests where a user was
querying why they issuer was failing as a jwks_uri isn't needed for VCI
however I'll fix that part in a separate MR.

There's probably more than could be done here - e.g. configuration fields
for server keys could be hidden when not needed - but this seems like a good
step in the right direction.