VCI Issuer: Handle authz server consistently in happy flow test
The happy flow test wasn't aligned with how the metadata test module
works - in particular it was allowing any authorization server to be
picked even if it wasn't in the issuer's list of authorization servers,
and it was fetching openid-configuration whereas VCI points people
towards the OAuth server metadata RFC.

The test user can now leave the discovery issuer box empty, and if left
empty the test should automatically work if the credential issuer's
metadata lists zero or 1 authorization servers. If more than one is
listed the user must specify in the test configuration which one they
want to use.

This could probably do with refactoring - reusing the test conditions
from the metadata test means that the metadata for all authorization
servers is being fetched, when it'd be better to fetch the one only for
the server under test.

The selected authorization server's data is placed into 'server' in the
environment so that all the existing test conditions just work.

This also corrects the issuer identifier that is returned in the mtls
case as mtls sender constraining/client authentication does not require
server metadata to be tls authenticated.